Create README.md
In this Project we took the opportunity to develop a reliable method for navigation inside buildings (usually big sanctuaries like libraries) where it’s often hard to reference oneself inside the building and where technologies like GPS fails to do so as well. As a medium, we use LEDs to send data, while on the other side, a camera will record the sent data in real time and reproduce the original data at the receiver side.